requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pyodbc==5.0.1
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.0.0
gunicorn==20.1.0
boto3==1.28.0
requests==2.31.0
